Day2/requirements.txt
numpy==1.23.0
Pillow==8.2.0